requirements.txt
torch>=2.0.0
sentencepiece>=0.1.99
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.28.0
numpy>=1.26.0,<2
beautifulsoup4>=4.12.0
lxml>=5.0.0
